requirements.txt
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.2.1
rsa==4.9.1
Telethon==1.42.0
